requirements.txt
absl-py==0.12.0
altgraph==0.17
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==4.1.0
certifi==2018.1.18
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
decorator==4.1.2
httplib2==0.9.2
idna==2.6
importlib-resources==5.1.4
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
oauth==1.0.1
oauthlib==3.2.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyserial==3.5
python-dateutil==2.6.1
python-dotenv==0.20.0
python3-xlib==0.15
PyTweening==1.0.3
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
reportlab==3.4.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
rise==5.7.1
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.13.2
six==1.11.0
systemd-python==234
testpath==0.6.0
tornado==6.1
traitlets==4.3.3
tweepy==4.6.0
urllib3==1.22
virtualenv==20.4.7
